+++ src/MemcachedConnector.php
@@ -2,7 +2,7 @@
 namespace SgIoc\Cache;
 /**
  *
- * User: Administrator
+ * User: freelife2020@163.com
  * Date: 2018/4/8
  * Time: 15:07
  */